+++ Projects/eLooM_Components/SensorManager/Inc/VD6283TXTask_vtbl.h
@@ -50,10 +50,10 @@ SensorDescriptor_t VD6283TXTask_vtblSensorGetDescription(ISensor_t *_this);
 SensorStatus_t VD6283TXTask_vtblSensorGetStatus(ISensor_t *_this);
 sys_error_code_t VD6283TXTask_vtblLightGetIntermeasurementTime(ISensorLight_t *_this, uint32_t *p_measured,
                                                                uint32_t *p_nominal);
-float VD6283TXTask_vtblLightGetExposureTime(ISensorLight_t *_this);
+uint32_t VD6283TXTask_vtblLightGetExposureTime(ISensorLight_t *_this);
 sys_error_code_t VD6283TXTask_vtblLightGetLightGain(ISensorLight_t *_this, float *LightGain);
 sys_error_code_t VD6283TXTask_vtblSensorSetIntermeasurementTime(ISensorLight_t *_this, uint32_t intermeasurement_time);
-sys_error_code_t VD6283TXTask_vtblSensorSetExposureTime(ISensorLight_t *_this, float exposure_time);
+sys_error_code_t VD6283TXTask_vtblSensorSetExposureTime(ISensorLight_t *_this, uint32_t exposure_time);
 sys_error_code_t VD6283TXTask_vtblSensorSetLightGain(ISensorLight_t *_this, float LightGain, uint8_t channel);
 
 #ifdef __cplusplus
